Add option to keep baked files, copy existing fix

+++ addon/operators/tlm.py
@@ -75,9 +75,11 @@ class TLM_CleanLightmaps(bpy.types.Operator):
 
         filepath = bpy.data.filepath
         dirpath = os.path.join(os.path.dirname(bpy.data.filepath), scene.TLM_EngineProperties.tlm_lightmap_savedir)
-        if os.path.isdir(dirpath):
-            for file in os.listdir(dirpath):
-                os.remove(os.path.join(dirpath + "/" + file))
+
+        if not bpy.context.scene.TLM_SceneProperties.tlm_keep_baked_files:
+            if os.path.isdir(dirpath):
+                for file in os.listdir(dirpath):
+                    os.remove(os.path.join(dirpath + "/" + file))
 
         for obj in bpy.context.scene.objects:
             if obj.type == 'MESH' and obj.name in bpy.context.view_layer.objects:
